Move fixture `populate_db_with_unprocessed_dataset_urls` to a `conftest.py`
