+++ datalad_registry/tests/test_new_organization/test_tasks/test_process_dataset_url.py
@@ -11,8 +11,6 @@ from sqlalchemy import inspect
 from datalad_registry.models import URL, db
 from datalad_registry.tasks import process_dataset_url
 
-_TEST_MIN_REPO_URL = "https://github.com/datalad/testrepo--minimalds.git"
-
 
 def is_there_file_in_tree(top: Union[Path, str]) -> bool:
     """
@@ -53,34 +51,6 @@ def is_dataset_url_unprocessed(dataset_url_id: int) -> bool:
     return (not dataset_url.processed) and all(
         getattr(dataset_url, f) is None for f in default_none_fields
     )
-
-
-@pytest.fixture
-def populate_db_with_unprocessed_dataset_urls(
-    flask_app,
-    empty_ds_annex,
-    empty_ds_non_annex,
-    two_files_ds_annex,
-    two_files_ds_non_annex,
-):
-    """
-    Populate the database with unprocessed dataset URLs
-    """
-
-    with flask_app.app_context():
-        non_dataset_url = URL(url="https://www.datalad.org/")
-        db.session.add(non_dataset_url)  # id == 1
-
-        db.session.add(URL(url=_TEST_MIN_REPO_URL))  # id == 2
-        db.session.add(URL(url=empty_ds_annex.path))  # id == 3
-        db.session.add(URL(url=empty_ds_non_annex.path))  # id == 4
-        db.session.add(URL(url=two_files_ds_annex.path))  # id == 5
-        db.session.add(URL(url=two_files_ds_non_annex.path))  # id == 6
-
-        db.session.commit()
-
-        db.session.delete(non_dataset_url)
-        db.session.commit()  # 1 is no longer a valid dataset URL id
 
 
 class TestProcessDatasetUrl:
